add exceptions to the vod module change the IDs datatype to Integer instead of Long

+++ domain/vod/src/main/java/entity/TvSeries.kt
@@ -2,25 +2,26 @@ package entity
 import kotlinx.datetime.LocalDate
 
 data class TvSeries(
-    val id: Long,
+    val id: Int,
     val posterImageUrl: String,
     val title: String,
     val description: String,
     val seasons: List<Season>,
-    val actorIds: List<Long>,
+    val actorIds: List<Int>,
     val releaseDate: LocalDate,
     val genres: List<Genre>,
     val imdbRating: Float?,
 )
+
 data class Season(
-    val id: Long,
+    val id: Int,
     val title: String,
     val number: Int,
     val episodes: List<Episode>
 )
 
 data class Episode(
-    val id: Long,
+    val id: Int,
     val title: String,
     val description: String,
     val durationMinutes: Int,
